requirements.txt
fastapi
uvicorn
black
sqlalchemy
passlib
bcrypt
python-jose